Move get_brain_fig
