+++ mne/viz/utils.py
@@ -2287,25 +2287,3 @@ def _ndarray_to_fig(img):
     fig = _figure_agg(dpi=dpi, figsize=figsize, frameon=False)
     fig.figimage(img, resize=True)
     return fig
-
-
-def get_brain_fig(brain):
-    """Return the figure associated to the given brain object.
-
-    Parameters
-    ----------
-    brain : object
-        The brain used to extract the figure.
-
-    Returns
-    -------
-    fig : object
-        The brain figure corresponding to the current 3d backend.
-    """
-    from .backends.renderer import get_3d_backend
-    if get_3d_backend() == "mayavi":
-        # from mayavi import mlab
-        # return mlab.gcf()
-        return brain.figures[0][0]
-    else:
-        return brain._renderer.scene()
